删除xm_sport.js，改用xmSports.js
