+++ package.json
@@ -50,7 +50,7 @@
     "bcs_checkin": "node other/bcs_checkin.js",
     "52pojie_checkin": "node other/52pojie_checkin.js",
     "ruike1_reply": "node other/ruike1_reply.js",
-    "xm_sport": "node xm_sport.js",
+    "xm_sport": "node backUp/xmSports.js",
     "qqread": "node backUp/qqread.js",
     "check_workflow": "node utils/check_workflow.js"
   },
